authentication and authorization MWs added

+++ Controllers/eventController.js
@@ -1,9 +1,38 @@
 const {validationResult} = require("express-validator");
 const Event = require("./../Models/eventModel");
 
+//Helper Functions
+function AssignRoleAdmin(req)
+{
+    //console.log(req.role);
+    if(req.role !== "admin")
+    {
+        throw new Error("Not Authorized");
+    }
+}
+
+function AssignRoleStudent(req)
+{
+    console.log(req.role)
+    if(req.role !== "student")
+    {
+        throw new Error("Not Authorized");
+    }
+}
+
+function AssignRoleSpeaker(req)
+{
+    if(req.role !== "speaker")
+    {
+        throw new Error("Not Authorized");
+    }
+}
+
+//Main Functions
 module.exports.getEvents = (req,res,next)=>{
 
-    events.find({})
+    AssignRoleAdmin(req);
+    Event.find({})
           .then((data)=>{
 
             res.status(200).json({data});
@@ -17,7 +46,17 @@ module.exports.getEvents = (req,res,next)=>{
 
 module.exports.getEventById = (req,res,next)=>{
 
-    events.find({_id:req.body.id})
+    AssignRoleAdmin(req);
+    //validation
+    let result = validationResult(req);
+    if (!result.isEmpty()) {
+        let message = result.array().reduce((current, error) => current + error.msg + ", ", " ");
+        let error = new Error(message);
+        error.status = 422;
+        throw error;
+    }
+    //response
+    Event.find({_id:req.params.id})
           .then((data)=>{
 
             res.status(200).json({data});
@@ -30,6 +69,8 @@ module.exports.getEventById = (req,res,next)=>{
 }
 
 module.exports.createEvent = (req,res,next)=>{
+
+    AssignRoleAdmin(req);
     //validation
     let result = validationResult(req);
     if(!result.isEmpty())
@@ -64,6 +105,17 @@ module.exports.createEvent = (req,res,next)=>{
 
 module.exports.updateEvent = (req,res)=>{
 
+    AssignRoleAdmin(req);
+     //validation
+     let result = validationResult(req);
+     if(!result.isEmpty())
+     {
+         let message = result.array().reduce((current,error)=>current + error.msg + ", " , " ");
+         let error = new Error(message);
+         error.status = 422;
+         throw error;
+     }
+     //response 
     Event.updateOne({_id:req.body.id},{
         $set:{
             _id : req.body.id,
@@ -78,9 +130,9 @@ module.exports.updateEvent = (req,res)=>{
 
                 if(data.matchedCount == 0)
                 {
-                    throw new Error("Speaker Not Found");
+                    throw new Error("Event Not Found");
                 }
-                res.status(200).json({message:"Speaker Updated",data});
+                res.status(200).json({message:"Event Updated",data});
 
            })
            .catch((error)=>{
@@ -92,6 +144,17 @@ module.exports.updateEvent = (req,res)=>{
 
 module.exports.deleteEvent = (req,res)=>{
 
+    AssignRoleAdmin(req);
+     //validation
+     let result = validationResult(req);
+     if(!result.isEmpty())
+     {
+         let message = result.array().reduce((current,error)=>current + error.msg + ", " , " ");
+         let error = new Error(message);
+         error.status = 422;
+         throw error;
+     }
+     //response 
     Event.deleteOne({_id:req.body.id})
            .then((data)=>{
 
@@ -106,4 +169,54 @@ module.exports.deleteEvent = (req,res)=>{
                next(error);
            })
     
+}
+
+module.exports.viewStudentEvents = (req,res,next)=>{
+
+    AssignRoleStudent(req);
+    //validation
+    let result = validationResult(req);
+    if(!result.isEmpty())
+    {
+        console.log(result);
+        let message = result.array().reduce((current,error)=>current + error.msg + ", " , " ");
+        let error = new Error(message);
+        error.status = 422;
+        throw error;
+    }
+    //response 
+    Event.find({students:req.params.id})
+    .then((data)=>{
+
+      res.status(200).json({data});
+
+    })
+    .catch((error)=>{
+        next(error);
+    })
+}
+
+module.exports.viewSpeakerEvents = (req,res,next)=>{
+
+    AssignRoleSpeaker(req);
+    //validation
+    let result = validationResult(req);
+    if(!result.isEmpty())
+    {
+        console.log(result);
+        let message = result.array().reduce((current,error)=>current + error.msg + ", " , " ");
+        let error = new Error(message);
+        error.status = 422;
+        throw error;
+    }
+    //response 
+    Event.find({$or: [ { mainSpeaker: req.params.id }, { otherSpeakers: req.params.id } ]})
+    .then((data)=>{
+
+      res.status(200).json({data});
+
+    })
+    .catch((error)=>{
+        next(error);
+    })
 }
