+++ Controllers/speakerController.js
@@ -16,6 +16,16 @@ module.exports.getSpeakers = (req,res,next)=>{
 
 module.exports.getSpeakerById = (req,res,next)=>{
 
+    //validation
+    let result = validationResult(req);
+    if (!result.isEmpty()) {
+        let message = result.array().reduce((current, error) => current + error.msg + ", ", " ");
+        let error = new Error(message);
+        error.status = 422;
+        throw error;
+
+    }
+    //response
     Speaker.find({_id:req.body.id})
     .then((data) => {
         res.status(200).json({ message: data });
@@ -67,6 +77,16 @@ module.exports.createSpeaker = (req,res,next)=>{
 
 module.exports.updateSpeaker = (req,res,next)=>{
 
+    //validation
+    let result = validationResult(req);
+    if (!result.isEmpty()) {
+        let message = result.array().reduce((current, error) => current + error.msg + ", ", " ");
+        let error = new Error(message);
+        error.status = 422;
+        throw error;
+
+    }
+    //response
     Speaker.updateOne({_id:req.body.id},{
         $set:{
             email : req.body.email,
@@ -98,6 +118,16 @@ module.exports.updateSpeaker = (req,res,next)=>{
 
 module.exports.deleteSpeaker = (req,res,next)=>{
     
+    //validation
+    let result = validationResult(req);
+    if (!result.isEmpty()) {
+        let message = result.array().reduce((current, error) => current + error.msg + ", ", " ");
+        let error = new Error(message);
+        error.status = 422;
+        throw error;
+
+    }
+    //response
     Speaker.deleteOne({_id:req.body.id})
            .then((data)=>{
